requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
mysql-connector-python>=8.0.26
python-dotenv>=0.19.0
bcrypt>=4.0.0